Extract error serializer
